Fixed several smaller issues caught by the compiler
Aside from a few miscelaneous unused variables, commented out -

In SDL2, SDL_Init returns a negative value for failure. In SDL3, SDLInit returns true for a success, and false for failure. Fixed the logic to accomodate this change.

Window sizes are now floats instead of ints, to play better with SDL3's FRects. (Avoids casting from int to float)

Re-ordered initialization of variables in tehBUS

Removed a few un-used variables from tehCPUS that were a left-over from my old error-handling code.

I was checking to see if an unsigned variable was greater than zero in I_FX0A_READ_KEY.

Removed some other cruft, and initialized some variables to sane values in tehVIDEO

+++ chipperSDL3.h
@@ -35,8 +35,8 @@ private:
     bool is_fullscreen;
     bool is_minimized;
     // Window dimensions
-    int window_height;
-    int window_width;
+    float window_height;
+    float window_width;
 
     // Variables used for tehBOOP
     SDL_Event input; // input queue
